@@ -10,8 +10,8 @@ except DistributionNotFound:
     pass
 
 __copyright__ = '''\
-© Crown Copyright 2017-2021 Defence Science and Technology Laboratory UK
-© Crown Copyright 2018-2021 Defence Research and Development Canada / Recherche et développement pour la défense Canada
-© Copyright 2018-2021 University of Liverpool UK
+© Crown Copyright 2017-2022 Defence Science and Technology Laboratory UK
+© Crown Copyright 2018-2022 Defence Research and Development Canada / Recherche et développement pour la défense Canada
+© Copyright 2018-2022 University of Liverpool UK
 '''  # noqa: E501
 __license__ = 'MIT'
