@@ -1,8 +1,12 @@
 # -*- coding: utf-8 -*-
+import numpy as np
 import pytest
 
+from ...predictor.categorical import HMMPredictor
 from ...types.prediction import (
-    GaussianMeasurementPrediction, GaussianStatePrediction)
+    GaussianMeasurementPrediction, GaussianStatePrediction, CategoricalStatePrediction,
+    CategoricalMeasurementPrediction)
+from ...updater.categorical import HMMUpdater
 
 
 @pytest.fixture()
@@ -23,3 +27,36 @@ def updater():
                                                  state_prediction.covar,
                                                  state_prediction.timestamp)
     return TestGaussianUpdater()
+
+
+@pytest.fixture()
+def dummy_category_predictor():
+    class DummyCategoricalPredictor(HMMPredictor):
+
+        @property
+        def transition_model(self):
+            pass
+
+        def predict(self, prior, timestamp=None, **kwargs):
+            """Return the same state vector."""
+            return CategoricalStatePrediction(prior.state_vector, timestamp=timestamp)
+
+    return DummyCategoricalPredictor()
+
+
+@pytest.fixture()
+def dummy_category_updater():
+    class DummyCategoricalUpdater(HMMUpdater):
+
+        @property
+        def measurement_model(self):
+            pass
+
+        def predict_measurement(self, state_prediction, measurement_model=None, **kwargs):
+            """Return the first two state vector elements, normalised."""
+            vector = state_prediction.state_vector[:2]
+            vector = vector / np.sum(vector)
+            return CategoricalMeasurementPrediction(vector,
+                                                    timestamp=state_prediction.timestamp)
+
+    return DummyCategoricalUpdater()
