@@ -51,11 +51,10 @@ def test_basicmetrics():
                                            end_timestamp=start_time +
                                            datetime.timedelta(seconds=4)),
                                        generator=generator)}
-
     for metric_name in ["Number of targets",
                         "Number of tracks", "Track-to-target ratio"]:
         calc_metric = [i for i in correct_metrics if i.title == metric_name][0]
-        meas_metric = [i for i in metrics if i.title == metric_name][0]
+        meas_metric = metrics.get(metric_name)
         assert calc_metric.value == meas_metric.value
         assert calc_metric.time_range.start_timestamp == \
             meas_metric.time_range.start_timestamp
