@@ -5,7 +5,7 @@ from ..base import Property
 from .array import CovarianceMatrix
 from .base import Type
 from .state import (State, GaussianState, ParticleState, SqrtGaussianState, InformationState,
-                    TaggedWeightedGaussianState, WeightedGaussianState)
+                    TaggedWeightedGaussianState, WeightedGaussianState, CategoricalState)
 from ..models.transition.base import TransitionModel
 
 
@@ -115,3 +115,11 @@ class ParticleMeasurementPrediction(MeasurementPrediction, ParticleState):
 
     This is a simple Particle measurement prediction object.
     """
+
+
+class CategoricalStatePrediction(Prediction, CategoricalState):
+    """Categorical state prediction type"""
+
+
+class CategoricalMeasurementPrediction(MeasurementPrediction, CategoricalState):
+    """Categorical measurement prediction type"""
