@@ -2,10 +2,9 @@
 import datetime
 
 import numpy as np
-import scipy.linalg
 import pytest
+import scipy.linalg
 
-from stonesoup.base import Property
 from stonesoup.types.groundtruth import GroundTruthState
 from stonesoup.types.state import CreatableFromState
 from ..angle import Bearing
@@ -13,7 +12,8 @@ from ..array import StateVector, CovarianceMatrix
 from ..numeric import Probability
 from ..particle import Particle
 from ..state import State, GaussianState, ParticleState, \
-    StateMutableSequence, WeightedGaussianState, SqrtGaussianState
+    StateMutableSequence, WeightedGaussianState, SqrtGaussianState, CategoricalState
+from ...base import Property
 
 
 def test_state():
@@ -431,3 +431,31 @@ def test_creatable_from_state_multi_base_error():
                        match='A CreatableFromState subclass must have exactly two superclasses'):
         class SubSubclassCfs(State, StateMutableSequence, SubclassCfs):
             pass
+
+
+def test_categorical_state():
+
+    # Test instantiation errors
+    with pytest.raises(ValueError, match="2 category names were given for a state vector with 3 "
+                                         "elements"):
+        CategoricalState(state_vector=StateVector([0.1, 0.4, 0.5]), category_names=['1', '2'])
+    with pytest.raises(ValueError, match=r"Category probabilities must lie in the closed interval "
+                                         r"\[0, 1\]"):
+        CategoricalState(state_vector=StateVector([1.1, 0.4, 0.5]))
+    with pytest.raises(ValueError, match="Category probabilities must sum to 1"):
+        CategoricalState(state_vector=StateVector([0.6, 0.7, 0.7]))
+    with pytest.raises(ValueError, match="Category probabilities must sum to 1"):
+        CategoricalState(state_vector=StateVector([0.2, 0.2, 0.3]))
+
+    # Test defaults
+    state = CategoricalState(state_vector=StateVector([0.1, 0.4, 0.5]))
+    assert state.category_names == [0, 1, 2]
+
+    # Test str
+    state = CategoricalState(state_vector=StateVector([0.1, 0.4, 0.5]),
+                             timestamp=datetime.datetime.now(),
+                             category_names=['red', 'blue', 'yellow'])
+    assert str(state) == "(P(red) = 0.1, P(blue) = 0.4, P(yellow) = 0.5)"
+
+    # Test category
+    assert state.category == 'yellow'
