@@ -4,7 +4,8 @@ from stonesoup.types.state import CreatableFromState
 from ..base import Property
 from .base import Type
 from .hypothesis import Hypothesis
-from .state import State, GaussianState, ParticleState, SqrtGaussianState, InformationState
+from .state import State, GaussianState, ParticleState, SqrtGaussianState,  \
+    InformationState, CategoricalState
 from .mixture import GaussianMixture
 
 
@@ -64,3 +65,7 @@ class InformationStateUpdate(Update, InformationState):
     This is a simple Information state update object, which, as the name
     suggests, is described by a precision matrix and its corresponding state vector.
     """
+
+
+class CategoricalStateUpdate(Update, CategoricalState):
+    """Categorical state prediction type"""
